Refactored the project
1. Fixed the pom dependencies
2. Fixed the package names
3. Removed JPA
4. Added Mybatis and postgres
