+++ src/main/java/com/educare/unitylend/ServletInitializer.java
@@ -1,0 +1,13 @@
+package com.educare.unitylend;
+
+import org.springframework.boot.builder.SpringApplicationBuilder;
+import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
+
+public class ServletInitializer extends SpringBootServletInitializer {
+
+    @Override
+    protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
+        return application.sources(UnitylendApplication.class);
+    }
+
+}
