+++ src/main/java/com/educare/unitylend/UnitylendApplication.java
@@ -1,0 +1,13 @@
+package com.educare.unitylend;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class UnitylendApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(UnitylendApplication.class, args);
+    }
+
+}
