+++ src/main/java/com/educare/unitylend/model/Community.java
@@ -1,0 +1,17 @@
+package com.educare.unitylend.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.UUID;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class Community {
+    private UUID communityId;
+    private String communityName;
+
+
+}
