+++ src/main/java/com/educare/unitylend/model/Wallet.java
@@ -1,0 +1,16 @@
+package com.educare.unitylend.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.UUID;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class Wallet {
+    private UUID Wallet;
+    private UUID userId;
+    private Float Balance;
+}
