+++ src/main/resources/application.properties
@@ -1,0 +1,10 @@
+
+server.port=8085
+# MY-batis postgres configuration
+#spring.datasource.url=jdbc:postgresql://adapter-rds-deshaw-poc.cckfesasu3h8.us-east-1.rds.amazonaws.com:5432/nautilus_deck
+#spring.datasource.username=genusr
+#spring.datasource.password=sometestpw
+spring.datasource.url=jdbc:postgresql://localhost:5432/unity-lend
+spring.datasource.username=user
+spring.datasource.password=password:q
+spring.datasource.driver-class-name=org.postgresql.Driver
