+++ src/test/java/com/educare/unitylend/UnityLendApplicationTests.java
@@ -1,0 +1,13 @@
+package com.educare.unitylend;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class UnityLendApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
