--- unitylend/src/main/java/com/group6/unitylend/ServletInitializer.java
@@ -1,13 +1,0 @@
-package com.group6.unitylend;
-
-import org.springframework.boot.builder.SpringApplicationBuilder;
-import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
-
-public class ServletInitializer extends SpringBootServletInitializer {
-
-	@Override
-	protected SpringApplicationBuilder configure(SpringApplicationBuilder application) {
-		return application.sources(UnitylendApplication.class);
-	}
-
-}
