--- unitylend/src/main/java/com/group6/unitylend/UnitylendApplication.java
@@ -1,13 +1,0 @@
-package com.group6.unitylend;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class UnitylendApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(UnitylendApplication.class, args);
-	}
-
-}
