--- unitylend/src/test/java/com/group6/unitylend/UnitylendApplicationTests.java
@@ -1,13 +1,0 @@
-package com.group6.unitylend;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class UnitylendApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
